Lazyのplugins整理
  { import = "user.lazy.plugins.colorschemes" },
  { import = "user.lazy.plugins.project" },
  { import = "user.lazy.plugins.status" },
  { import = "user.lazy.plugins.edit" },
  { import = "user.lazy.plugins.filer" },
  { import = "user.lazy.plugins.terminal" },
  { import = "user.lazy.plugins.keymap" },
  { import = "user.lazy.plugins.cmp" },
