+++ init.lua
@@ -1,27 +1,10 @@
 require("user.options")
-require("user.lazy")
 require("user.keymaps")
+require("user.lazy")
 require("user.colorscheme")
 require("user.autocommands")
 
 -- the others
 require("user.yoshi")
-
--- move to config in plugins.lua
--- require("user.toggleterm")
--- require("user.whichkey")
--- require("user.cmp")
--- require("user.lsp")
--- require("user.telescope")
--- require("user.treesitter")
--- require("user.autopairs")
--- require("user.comment")
--- require("user.gitsigns")
--- require("user.nvim-tree")
--- require("user.bufferline")
--- require("user.lualine")
--- require("user.impatient")
--- require("user.indentline")
--- require("user.alpha")
 
 vim.g.dbext_default_ORA_bin = "sql"
