+++ lua/user/lazy/plugins/cmp/friendly-snippets.lua
@@ -1,0 +1,4 @@
+-- a bunch of snippets to use
+return {
+  "rafamadriz/friendly-snippets",
+}
