+++ lua/user/lazy/plugins/cmp/luasnip.lua
@@ -1,0 +1,7 @@
+--snippet engine
+return {
+  "L3MON4D3/LuaSnip",
+  config = function()
+    require("luasnip.loaders.from_vscode").lazy_load({ paths = { "./snippets" } })
+  end,
+}
