+++ lua/user/lazy/plugins/colorschemes/darkplus.lua
@@ -1,0 +1,3 @@
+return {
+  "lunarvim/darkplus.nvim",
+}
