+++ lua/user/lazy/plugins/colorschemes/dracula.lua
@@ -1,0 +1,3 @@
+return {
+  "Mofiqul/dracula.nvim"
+}
