+++ lua/user/lazy/plugins/colorschemes/tokyonight.lua
@@ -1,0 +1,3 @@
+return {
+  "folke/tokyonight.nvim", opt = true,
+}
