+++ lua/user/lazy/plugins/edit/comment.lua
@@ -1,0 +1,5 @@
+return {
+  "numToStr/Comment.nvim",
+  event = { "VimEnter" },
+  opts = {},
+}
