+++ lua/user/lazy/plugins/edit/nvim-ts-context-commentstring.lua
@@ -1,0 +1,4 @@
+return {
+  "JoosepAlviste/nvim-ts-context-commentstring",
+  event = { "VimEnter" },
+}
