+++ lua/user/lazy/plugins/project/project.lua
@@ -1,0 +1,5 @@
+-- project.nvimはgitのプロジェクト配下のどこで開いてもgit rootに自動でcdしてくれるプラグイン
+return {
+  "ahmedkhalf/project.nvim",
+  event = { "BufRead", "BufNewFile" },
+}
