+++ lua/user/lazy/plugins/status/lualine.lua
@@ -1,8 +1,3 @@
-local status_ok, lualine = pcall(require, "lualine")
-if not status_ok then
-  return
-end
-
 local hide_in_width = function()
   return vim.fn.winwidth(0) > 80
 end
@@ -74,7 +69,7 @@ local spaces = function()
   return "spaces: " .. vim.api.nvim_get_option_value("shiftwidth", { buf = 0 })
 end
 
-lualine.setup({
+local opts = {
   options = {
     icons_enabled = true,
     theme = "auto",
@@ -105,4 +100,10 @@ lualine.setup({
   },
   tabline = {},
   extensions = {},
-})
+}
+
+
+return {
+  "nvim-lualine/lualine.nvim",
+  opts = opts,
+}
